svc: add DbStructureTest
It is adobted implementation of tests created initally
to verify db subproject which was removed then.

Style is kept almost as is so looks a bit weird comparing to other code.

The reason is that the tests were created at very beginning
when Scala experience was quite poor.

See no reason though to put too much effort to rework
since they do what the should do.

It completes db subproject removal.